requirements.txt
blinkpy
git+https://github.com/Rapptz/discord.py.git
git+https://github.com/jasonacox/tinytuya.git

python-dotenv==0.19.0